update 1 dong api
