+++ tennisTournament/tennis/src/main/java/com/example/demo/api/RoundApi.java
@@ -1,6 +1,5 @@
 package com.example.demo.api;
 
-import com.example.demo.entity.Match;
 import com.example.demo.entity.Round;
 import com.example.demo.exception.ResourceNotFoundException;
 import com.example.demo.service.RoundService;
@@ -9,37 +8,40 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.HashMap;
+import java.net.URI;
 import java.util.List;
-import java.util.Map;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/round")
+@RequestMapping(RoundApi.PATH)
 public class RoundApi {
     @Autowired
-    private final RoundService roundService;
+    private RoundService roundService;
+
+    public static final String PATH = "/api/round";
 
     @GetMapping
-    public List<Round> getAll(){
-        return roundService.getAll();
+    public ResponseEntity<List<Round>> getAll(){
+        return ResponseEntity.ok(roundService.getAll());
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<Round> getById(@PathVariable(value = "id")Integer id) throws ResourceNotFoundException {
         Round round = roundService.findRoundById(id).orElseThrow(() -> new ResourceNotFoundException("Id not found:" + id));
-        return ResponseEntity.ok().body(round);
+        return ResponseEntity.ok(round);
     }
 
-    @PostMapping("/add")
-    public Round create(@RequestBody Round round){
-        return roundService.saveRound(round);
+    @PostMapping
+    public ResponseEntity<Round> create(@RequestBody Round round){
+        Round createdRound = roundService.saveRound(round);
+        return ResponseEntity.created((URI.create(RoundApi.PATH + "/" + createdRound.getRoundName()))).body(createdRound);
     }
 
     @PutMapping("/{id}")
     public ResponseEntity<Round> update(@PathVariable(value = "id")Integer id,
                                         @RequestBody Round roundDetail) throws ResourceNotFoundException{
         Round round = roundService.findRoundById(id)
+
                 .orElseThrow(() -> new ResourceNotFoundException("Id not found: " + id));
         round.setRoundName(roundDetail.getRoundName());
         round.setRoundType(roundDetail.getRoundType());
@@ -49,12 +51,12 @@ public class RoundApi {
     }
 
     @DeleteMapping("/{id}")
-    public Map<String, Boolean> delete(@PathVariable(value = "id")Integer id) throws ResourceNotFoundException{
+    public ResponseEntity<Void> delete(@PathVariable(value = "id")Integer id) throws ResourceNotFoundException{
         Round round = roundService.findRoundById(id)
+
                 .orElseThrow(() -> new ResourceNotFoundException("Id not found: " + id));
-        roundService.deleteRoundById(id);
-        Map<String,Boolean> response = new HashMap<>();
-        response.put("delete", Boolean.TRUE);
-        return response;
+        roundService.deleteRoundById(id)
+        ;
+        return ResponseEntity.noContent().build();
     }
 }
